+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * System
   * Fixed a bug where support tunnel addresses would be served when querying for amplipi.local
   * Stream validation for URLs has been made more robust
+  * Minor bugfixes with Airplay 1 stream handling
 
 ## 0.4.2
 * Streams